backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pymongo==4.3.3
python-dotenv==1.0.0
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.2
joblib==1.2.0
twilio==8.1.0
PyPDF2==3.0.1
